Initial foray at integration embedded PostgreSQL.
Initially only applies for #[cfg(test)] types of targets, to
convert "integration tests" into "unit tests".

The special system constructor returns a handle to the temporary
database process, which *must be held onto* through the duration
of a given test, as having it Drop'd causes the DB to shutdown.

Note: this includes assigning it to _ or _db, hence using an
actually named (but other unused) variable.