feat: support AND/OR operators and full-text search
The q param is assumed to be a url-encoded string of the form,
`{text}(&{field}{op}{value})*`

The {text} portion will become an OR clause of LIKE operators on all
the string fields in the entity. The optional AND filters will further
constrain the results.

Because the {text} can be anything, previously invalid filters became
valid.

We probably need to support escaping ampersands and pipes at some
point.

Signed-off-by: Jim Crossley <jim@crossleys.org>