refactor: switch from env_logger to tracing logger
Instead of pushing tracing events to log, and from log to env_logger,
this consumes log into tracing, and then logs those with trace spans
in mind.

This mostly means that trace spans only get logged when they fail.

It still uses the RUST_LOG variable, and defaults to INFO.

On top of that, there's color.