Move raw sbom/advisory endpoints beneath storage module
Fixes #255

We fetch them by their digest only, since database id's aren't exposed
to users. It's worth noting that the implementations for returning
either sbom or advisory are virtually identical, since storage only
cares about a digest, no matter what kind of doc it refers to.

So we might just consider replacing these with a single
`/api/v1/download/{sha256}` api. That way, we'd avoid having to do
this...

We removed the web::scope's for sbom's and advisories in the fetch
module because having them seemed to prevent the storage modules
endpoints from being configured. I'm not exactly sure why.

Signed-off-by: Jim Crossley <jim@crossleys.org>