feat: new convention for organizing integration tests
In the interest of both more helpful output and faster build times, I
propose we keep only a single rust file at the root of each module's
`tests/` directory representing a crate of integration tests for that
module. Its name should match that of the module, and it will load all
its tests from its own submodules.

Because each rust file beneath `tests/` is built as a crate rather
than a module by the test framework, we don't get the benefit of
parallel test execution if there is more than one file, i.e. crate, in
the root of `tests/`.

The naming convention is so we'll see similar to the following in the
build output, easily distinguishing unit and integration tests.

     Running unittests src/lib.rs (target/debug/deps/trustify_module_ingestor-006fc6101daf6771)
...
     Running tests/ingestor.rs (target/debug/deps/ingestor-1c5097ecb9d08752)
...

Signed-off-by: Jim Crossley <jim@crossleys.org>