refactor: simplify tests by incorporating more TrustifyContext
Expanded CallService to return json and added a `caller` helper to
return an impl of it, i.e. an App::new configured with the fundamental
endpoints. Replaced the `with_upload` async closures with calls to
`caller` and `ctx.ingest_document`. Added an explicit test for SBOM
uploading in lieu of `with_upload`.

Signed-off-by: Jim Crossley <jim@crossleys.org>