feat: replace SyncAdapter with SyncIoBridge to reduce RAM usage
Because serde deserialization is synchronous, and because we
immediately read back the bytes we store when ingesting a doc (to
ensure we can), we have to store them in RAM before asking serde to
deserialize them.

So now we deserialize through the bridge, hopefully using less RAM in
the process.

Signed-off-by: Jim Crossley <jim@crossleys.org>