build: update Rust to 1.80.1
Required by https://github.com/trustification/trustify/issues/703

This also requires us to drop the "mock" feature cfg. It currently is
an undeclared feature, which is a warning now, and might become an error
later.

Declaring that feature and mapping it to `sea-orm/mock` would be the
right thing to do. However, sea-orm will drop the `Clone` implementation
of the connection trait in the case the mock feature is enabled. And
our database struct requires the connection trait to be `Clone`.
Meaning, that we don't support the `mock` feature anyway.