refactor: use GitWalker directly instead of the wrappers
The wrappers seemed to be mostly collections of methods that did
little more than delegate to their GitWalker field, but the essential
complexity distinguishing each wrapper was in their handler impl, so
using the GitWalker directly seems simpler.

This makes it much easier to add behavior to all the wrappers,
e.g. setting the `depth` fetched from the git repo.

Signed-off-by: Jim Crossley <jim@crossleys.org>