chore: reduce the amount of data
Right now, the "fetch sbom (information)" call generates all kind of
details, which in most cases will be discarded. However, generating
this kind of information takes time, and also produces an enormous
amount of data in the response (cases of ~70MiB).

This PR splits up the call into two: summary and details. It also
applies pagination to the advisories of the details.