refactor: use the actual openapi spec, instead of a constructed one
By leveraging utoipa_actix_web, this change removes the hand-crafted
collection of openapi operations, and let utoipa collect this during
the creation of the actix web `App`. This results in the actual spec,
and removes the need to keep this aligned with the `openapi` functions
we had.

The downside is, that the `xtask` now needs to construct a full `App`
instance. Which isn't a big deal, but requires an embedded database
to the instantiated. However, we do this for running tests too.

This also already fixed an missing operation.