Add a means to alias query-able table names
The `find_also_linked` SeaORM method aliases a join table to `r0`. I
would've thought postgres would allow referring to either the actual
name or the alias, but apparently only the alias is allowed. So we now
have a method to do that on the `Columns` struct.

This ultimately enables the `sbom_node.name` column to be
queried/sorted in the `fetch_sboms` fn.

Signed-off-by: Jim Crossley <jim@crossleys.org>