fix: ensure we always detect a duplicate source document
Having multiple unique indices on the source document table, it might
happen that the sha256 is not the first one being reported. This may
lead to the issue that the "on conflict" statement will not be respected
and that we still get an internal error reported. As, right now, it is
not possible to use "on conflict" with multiple constraints
independently, we need to manually check for this error and also use
nested transactions, in order to continue with the transaction.

Closes: #1405