fix: purl's/cpe's now query-able via /api/v2/analysis/components
Fixes: #1280

This required translating CPE queries to their constituent fields in
the CPE table.

We converted query::Value from an enum to a trait to query both purl's
and cpe's because simple string comparisons won't cut it for both the
loading of relevant objects *and* the comparisons within the analysis
graph.

So we delegate the comparison logic to the disparate objects being
queried -- each must now implement the Value trait and corresponding
trait objects are then used to find matches.