fix: prevent duplicate PURLs or CPEs being returned
The query of encoding a list of PURLs (and CPEs) into a single field
did create duplicates. This is now properly being aggregated. One
requirement of this the ability to compare PURLs, which isn't possible
using the JSON type in PSQL. It works with strings or rows though.

Therefore, and also because it was overdue, this PR switches to
extracting the PURL from the `purl` field, instead of the three layer
table structure. Which might have a positive performance effect too.

Closes: #1417