feat: filter components by partial matches of a purl and/or cpe
Fixes: #1280

Added the proper logic to translate CPE queries into their constituent
columns and Purl queries into comparisons against the get_purl() proc,
which stringifies the associated JSON blob in the database.

This accounts for the loading of the graph. The fact that both Cpe and
Purl structs now implement Valuable accounts for querying the graph,
i.e. applying the query to a context constructed from the graph.