fix: prevent an OOM situation by using a stream as result
Using a vec requires a lot of entries to be held in memory. In one
example, around 1.6 million. Having multiple, parallel queries like
that can trigger an out of memory situation.

Using a stream for processing the result resolves this, and also allows
more await points interrupting the processing, as it's waiting on the
result from the database, during processing.