fix: prevent deadlock when inserting in parallel
Holding the lock for the whole time of the loop seems causing
deadlocks when other parts of the loop block. For example, creating an
organization might block.

In that case it may be that task A holds the org lock and waits for the
db context, while task B holds the db context and waits or the org lock.

It should be fine just holding the db context lock for the get call.

Closes: #1531