refactor: replace Operand/Arg structs with SeaOrm equivalent
This is a pure refactoring, even though one of the filtering tests
that relied on knowing the internal structure of Filter was
eliminated. The behavior it affirmed was already affirmed in other
tests.

This reflects a better understanding of the SeaORM abstractions and
allows us to better support more custom SimpleExpr's involving not
only the column type but the value queried.