feat: support unlimited multi-part field names for json columns
We no longer limit field names to having only one `:` delimiting the
json column name and its top-level key. As many delimiters as are
required to query nested objects are allowed.

This impacts the sort logic a bit, as the last field delimited by ':'
is used to specify direction, either 'asc' or 'desc'.

This change only affects SQL queries. The application of queries
containing deeply nested fields on in-memory values will be addressed
in a subsequent commit.