fix: set vulnerability score from cve advisory
This commit changes how vulnerability scoring is set. Instead of calculating an average
from all advisories, it is now parsed from the CVE file. The score and severity values
are stored in the vulnerability database.

BREAKING-CHANGE: The /vulnerabilities search paramteres are changed from 'average_severity'
and 'average_score' to 'base_severity' and 'base_score'