feat: import SBOM attachments from Quay repositories
May be configured with a registry source, a namespace and a size
limit. The only namespace I've found any SBOM's in is
'redhat-user-workloads' which contains ~150K SBOM's within ~8K repos.

Quay API requests are concurrently fetched, up to 32 at a time, but
ingestions occur synchronously, so as not to overwhelm the DB.

The Quay API token should be available in the QUAY_API_TOKEN
environment variable. The importer will work without it, but probably
won't find much.