refactor: generate lowercase enum variants, accept all cases
Currently, the enums variants supported as string are camelCase, which
results in lowercase for acronyms when using strum. However, when using
serde, which is used for generating the schema, it results in weird
cases like `oSV`.

This change aligns this, so that both strum as well as the schema use
the same, lowercase variants.