refactor: generate lowercase enum variants, accept all cases
Currently, the enums variants supported as string are camelCase, which
results in lowercase for acronyms when using strum. However, when using
serde, which is used for generating the schema, it results in weird
cases like `oSV`.

This change aligns this, so that both strum as well as the schema use
the same, lowercase variants.

(cherry picked from commit 7d02bcff9357182d7cc3cafc61e2a7247e0a0ca1)