feat: start exctracting vulnerability scores from cvss v4 and v2
Currently we only extract base score and severity values. There's need to be more work done
to properly save all scores of all versions for vulnerabilities and advisories.
The base score used for the vulnerability will be picked from the highest cvss version
available. Since base severity was not standardized in cvss v2.0 it is recalculated from
the score.