fix: align the embedded postgres version with the CI
In PM mode, the postgresql binary will be embedded. This requires the
build to know which version to download. However, this didn't align
with the code requesting a database.

This change prefers the use of the CI version, but falls back to the
locally (code) provided value. This ensures that if the CI builds with
a provided version, it will be used during runtime.

Closes: #1674
(cherry picked from commit 50e731d7993c9fd5e81fa0b052b569e4516f90dc)