feat(analysis): ensure parallel loads await each other
Loading the same SBOM twice is currently rather inefficient. As the work
is done twice (nor more times), but then discarded in the final step.

This PR changes this in a way a load operation is registered and
following cache misses will await that operation, rather than starting
a new one.