fix(fundamental): change response of analyse purl endpoint
Skip returning SBOM specific content as the endpoint should return
results for this PURL, not related SBOMs.

Remove duplicates in response. Due to the way the SQL query was
structured, it returned duplicate entries.

This will now return all advisories contributing to this state. It will
also skip not_affected and recommended, in addition to fixed.

Maps will also have a stable, sorted order of keys.