fix: prevent calculating the total if we don't need it
In some cases we calculate a "total number of items" although we don't
need it. This is due to the fact that we support pagination, and for
this we need to execute the SQL statement twice: once for getting
the page of items, and once for getting the total number of items.

In some cases however, we just drop the total, without ever using it. In
such cases executing the count() statement is a wast of time and
resources.

This change allows the caller of a function to influence the output type
and with that trigger an executing of the count() variant, or skip it.