refactor: handle read-only errors as "unavailable"
This change handles database errors of the kind "read-only" when
mutating as an "unavailable" error, rather than an internal server
error.

By default this doesn't change anything, as all transactions are
read-write by default. However, it now is possible to set the database
into a "read-only" by default mode, and have trustify handle this more
gracefully.